requirements.txt
bootstrap4==0.1.0
certifi==2019.6.16
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.8
django-channels==0.7.0
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
whitenoise==4.1.2
